+++ src/Points.php
@@ -5,6 +5,7 @@ namespace PHPWorldWide\Stats;
 use PHPWorldWide\Stats\Model\Topic;
 use PHPWorldWide\Stats\Model\Comment;
 use PHPWorldWide\Stats\Model\Reply;
+use PHPWorldWide\Stats\Util\CodeDetector;
 
 /**
  * Class Points.
@@ -12,24 +13,69 @@ use PHPWorldWide\Stats\Model\Reply;
 class Points
 {
     /**
-     * @var Config
-     */
-    private $config;
-
-    /**
      * @var ExpressionLanguage
      */
     private $language;
 
     /**
-     * Points constructor.
-     *
-     * @param Config $config
+     * @var CodeDetector
      */
-    public function __construct(Config $config)
+    private $codeDetector;
+
+    /**
+     * @var array
+     */
+    private $points;
+
+    /**
+     * @var array
+     */
+    private $admins;
+
+    /**
+     * @var array
+     */
+    private $offensiveWords;
+
+    /**
+     * Points constructor.
+     */
+    public function __construct()
     {
-        $this->config = $config;
         $this->language = new ExpressionLanguage();
+        $this->codeDetector = new CodeDetector();
+    }
+
+    /**
+     * Set points configuration values.
+     *
+     * @param array $points
+     */
+    public function setPoints($points)
+    {
+        $this->points = $points;
+        $this->codeDetector->setPatterns($points['code_regex']);
+        $this->codeDetector->setMinCodeLines($points['min_code_lines']);
+    }
+
+    /**
+     * Set admins from configuration values.
+     *
+     * @param array $admins
+     */
+    public function setAdmins($admins)
+    {
+        $this->admins = $admins;
+    }
+
+    /**
+     * Set offensive words.
+     *
+     * @param array $offensiveWords
+     */
+    public function setOffensiveWords($offensiveWords)
+    {
+        $this->offensiveWords = $offensiveWords;
     }
 
     /**
@@ -42,25 +88,28 @@ class Points
     public function addPointsForTopic(Topic $topic)
     {
         // Add points for creating a topic
-        $points = $this->config->get('points_for_topic');
+        $points = $this->points['topic'];
 
         // Add points based on the topic reactions
-        $points += $this->language->evaluate($this->config->get('points_for_topic_reactions'), ['topic' => $topic]);
+        $points += $this->language->evaluate($this->points['topic_reactions'], ['topic' => $topic]);
 
         // Add points for using recommended links
         $points += $this->addPointsForLinks($topic->getMessage());
 
         // Set points for only photo shares
-        $points = $this->language->evaluate($this->config->get('points_for_image'), ['topic' => $topic, 'points' => $points]);
+        $points = $this->language->evaluate($this->points['image'], ['topic' => $topic, 'points' => $points]);
 
         // Set points if topic is closed for comments
-        $points = $this->language->evaluate($this->config->get('points_for_closed_topic'), ['topic' => $topic, 'points' => $points]);
+        $points = $this->language->evaluate($this->points['closed_topic'], ['topic' => $topic, 'points' => $points]);
+
+        // Set points if message contains code
+        $points = $this->language->evaluate($this->points['code'], ['codeDetector' => $this->codeDetector, 'message' => $topic->getMessage(), 'points' => $points]);
 
         // Remove points for using offensive words
         $points += $this->getOffensivePoints($topic->getMessage());
 
         // Add points for special admin topics
-        $points += $this->language->evaluate($this->config->get('points_for_admin_topics'), ['topic' => $topic, 'admins' => $this->config->getParameter('admins')]);
+        $points += $this->language->evaluate($this->points['admin_topic'], ['topic' => $topic, 'admins' => $this->admins]);
 
         return $points;
     }
@@ -74,12 +123,15 @@ class Points
      */
     public function addPointsForComment(Comment $comment)
     {
-        $points = $this->config->get('points_for_comment');
+        $points = $this->points['comment'];
         $points += $this->getPointsForLikes($comment->getLikesCount());
         $points += $this->getPointsForMessageLength($comment->getMessage());
 
         // Add points for using recommended links
         $points += $this->addPointsForLinks($comment->getMessage());
+
+        // Set points if message contains code
+        $points = $this->language->evaluate($this->points['code'], ['codeDetector' => $this->codeDetector, 'message' => $comment->getMessage(), 'points' => $points]);
 
         // Remove points for using offensive words
         $points += $this->getOffensivePoints($comment->getMessage());
@@ -96,12 +148,15 @@ class Points
      */
     public function addPointsForReply(Reply $reply)
     {
-        $points = $this->config->get('points_for_reply');
+        $points = $this->points['reply'];
         $points += $this->getPointsForLikes($reply->getLikesCount());
         $points += $this->getPointsForMessageLength($reply->getMessage());
 
         // Add points for using recommended links
         $points += $this->addPointsForLinks($reply->getMessage());
+
+        // Set points if message contains code
+        $points = $this->language->evaluate($this->points['code'], ['codeDetector' => $this->codeDetector, 'message' => $reply->getMessage(), 'points' => $points]);
 
         // Remove points for using offensive words
         $points += $this->getOffensivePoints($reply->getMessage());
@@ -122,7 +177,7 @@ class Points
         $negativePoints = 0;
 
         if (isset($message)) {
-            foreach ($this->config->get('urls') as $url) {
+            foreach ($this->points['urls'] as $url) {
                 if (false !== stripos($message, $url[0])) {
                     $positivePoints = ($url[1] > 0 && $positivePoints < $url[1]) ? $positivePoints : $url[1];
                     $negativePoints = ($url[1] <= 0 && $negativePoints > $url[1]) ? $negativePoints : $url[1];
@@ -144,7 +199,7 @@ class Points
     {
         $points = 0;
         if (isset($message)) {
-            foreach ($this->config->get('offensive_words') as $keyword) {
+            foreach ($this->offensiveWords as $keyword) {
                 if (false !== stripos($message, str_rot13($keyword[0]))) {
                     $points = ($points < $keyword[1]) ? $points : $keyword[1];
                 }
@@ -163,7 +218,7 @@ class Points
      */
     private function getPointsForMessageLength($message)
     {
-        return $this->language->evaluate($this->config->get('points_message_length'), ['message' => $message]);
+        return $this->language->evaluate($this->points['message_length'], ['message' => $message]);
     }
 
     /**
@@ -175,6 +230,6 @@ class Points
      */
     private function getPointsForLikes($likes)
     {
-        return $this->language->evaluate($this->config->get('points_for_likes'), ['likes' => $likes]);
+        return $this->language->evaluate($this->points['likes'], ['likes' => $likes]);
     }
 }